refactor: clean up subscription controller
